# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BetRoyal Betting & Casino Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /bets:
    post:
      operationId: placeBet
      tags: [bets]
      summary: Place a bet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
    get:
      operationId: getBets
      tags: [bets]
      summary: Get bet history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetList"
...
  /bonuses/{id}/claim:
    post:
      operationId: claimBonus
      tags: [bonuses]
      summary: Claim a bonus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBonus"

  # REFERRALS
...
  /admin/bonuses:
    get:
      operationId: getAdminBonuses
      tags: [admin]
      summary: Get all bonuses (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bonus"
    post:
      operationId: createAdminBonus
      tags: [admin]
      summary: Create a bonus (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BonusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bonus"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, role, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, agent]
        status:
          type: string
          enum: [active, suspended, banned]
        avatar:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        vipLevel:
          type: integer
        referralCode:
          type: string
        createdAt:
          type: string
...
    BetInput:
      type: object
      required: [gameId, amount, gameData]
      properties:
        gameId:
          type: integer
        amount:
          type: number
          minimum: 0.1
        gameData:
          type: object
    Bet:
      type: object
      required: [id, userId, gameId, amount, payout, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        gameId:
          type: integer
        gameName:
          type: string
        amount:
          type: number
        payout:
          type: number
        multiplier:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, won, lost, cancelled]
        gameData:
          type: object
        createdAt:
          type: string
    BetList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Bet"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Bonus:
      type: object
      required: [id, name, type, value, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [welcome, first_deposit, cashback, daily, vip, referral]
        value:
          type: number
        valueType:
          type: string
          enum: [percentage, fixed]
        minDeposit:
          type: ["number", "null"]
        maxBonus:
          type: ["number", "null"]
        wagerRequirement:
          type: number
        expiresIn:
          type: ["integer", "null"]
        isActive:
          type: boolean
    UserBonus:
      type: object
      required: [id, bonusId, userId, amount, wagerRequired, wagerCompleted, status, createdAt]
      properties:
        id:
          type: integer
        bonusId:
          type: integer
        userId:
          type: integer
        bonus:
          $ref: "#/components/schemas/Bonus"
        amount:
          type: number
        wagerRequired:
          type: number
        wagerCompleted:
          type: number
        status:
          type: string
          enum: [active, completed, expired, cancelled]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    BonusInput:
      type: object
      required: [name, type, value, valueType, wagerRequirement]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        value:
          type: number
        valueType:
          type: string
        minDeposit:
          type: number
        maxBonus:
          type: number
        wagerRequirement:
          type: number
        expiresIn:
          type: integer
        isActive:
          type: boolean

    # REFERRALS
...